fix: [SRTP-1753] add test util to reduce code duplication
